+++ easycv/models/detection/detr/detr_head.py
@@ -5,7 +5,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from easycv.models.builder import HEADS, build_transformer
-from easycv.models.detection.utils import (HungarianMatcher, accuracy,
+from easycv.models.detection.utils import (MLP, HungarianMatcher, accuracy,
                                            box_cxcywh_to_xyxy,
                                            box_xyxy_to_cxcywh,
                                            generalized_box_iou)
@@ -187,22 +187,6 @@ class PostProcess(nn.Module):
         }
 
         return results
-
-
-class MLP(nn.Module):
-    """ Very simple multi-layer perceptron (also called FFN)"""
-
-    def __init__(self, input_dim, hidden_dim, output_dim, num_layers):
-        super().__init__()
-        self.num_layers = num_layers
-        h = [hidden_dim] * (num_layers - 1)
-        self.layers = nn.ModuleList(
-            nn.Linear(n, k) for n, k in zip([input_dim] + h, h + [output_dim]))
-
-    def forward(self, x):
-        for i, layer in enumerate(self.layers):
-            x = F.relu(layer(x)) if i < self.num_layers - 1 else layer(x)
-        return x
 
 
 class SetCriterion(nn.Module):
