+++ easycv/models/detection/utils/__init__.py
@@ -7,3 +7,4 @@ from .generator import MlvlPointGenerator
 from .matcher import HungarianMatcher
 from .misc import (accuracy, filter_scores_and_topk, fp16_clamp, interpolate,
                    multi_apply, output_postprocess, select_single_mlvl)
+from .transformer import MLP
